Use Prometheus service discovery instead of static config

+++ Makefile
@@ -1,7 +1,7 @@
 .PHONY: build
 
 build:
-	curl -o bin/rq https://s3-eu-west-1.amazonaws.com/record-query/record-query/x86_64-unknown-linux-musl/rq
+	test -f bin/rq || curl -o bin/rq https://s3-eu-west-1.amazonaws.com/record-query/record-query/x86_64-unknown-linux-musl/rq
 	chmod +x bin/*
 	docker build -t philipnrmn/prometheus-on-dcos .
